fix:preserve ai chat history when switching views
